+++ app/models/payments/bank_transfer_payment.rb
@@ -1,0 +1,2 @@
+class BankTransferPayment < Payment
+end