+++ app/models/payments/paypal_payment.rb
@@ -1,0 +1,2 @@
+class PaypalPayment < Payment
+end