+++ db/migrate/20200510142507_add_type_to_payment.rb
@@ -1,0 +1,6 @@
+class AddTypeToPayment < ActiveRecord::Migration[6.0]
+  def change
+    add_column :payments, :type, :integer, limit: 1
+    add_column :payments, :external_service_response, :json, default: {}
+  end
+end