+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_04_11_174503) do
+ActiveRecord::Schema.define(version: 2020_05_10_142507) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -200,6 +200,8 @@ ActiveRecord::Schema.define(version: 2020_04_11_174503) do
     t.integer "amount_cents"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.integer "type", limit: 2
+    t.json "external_service_response", default: {}
     t.index ["order_id"], name: "index_payments_on_order_id"
     t.index ["organization_id"], name: "index_payments_on_organization_id"
     t.index ["user_id"], name: "index_payments_on_user_id"